requirements.txt
python-pptx>=0.6.21
markdown>=3.4.3
python-frontmatter>=1.0.0
PyYAML>=6.0.1
python-dotenv>=1.0.0 